requirements.txt
# 核心实验依赖：LoRA 训练、数据处理、评估与配置读取
torch>=2.3.0
transformers>=4.45.0
datasets>=2.20.0
peft>=0.12.0
accelerate>=0.33.0
evaluate>=0.4.2
pyyaml>=6.0.1
numpy>=1.26.0
tqdm>=4.66.0
scikit-learn>=1.4.0
jsonlines>=4.0.0
pytest>=8.0.0
matplotlib>=3.8.0
